% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optim_thres.R
\name{apply_optim}
\alias{apply_optim}
\title{Check and apply optimization for method}
\usage{
apply_optim(data, input_cols, method = NULL)
}
\arguments{
\item{data}{Data frame containing leaf fresh water mass and leaf water potentials}

\item{input_cols}{A vector of column names or indices for the data frame. The second, third, and fourth elements should correspond to fresh weight, water potential, and dry mass respectively.}

\item{method}{Method to optimize the number of points above and below turgor loss point. Options are "rmse", "aicc", or "r2".}
}
\value{
A list containing the number of points above and below turgor loss point.
}
\description{
Check and apply optimization for method
}
