% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clipped_lm.R
\name{autoplot.clipped_lm_fit}
\alias{autoplot.clipped_lm_fit}
\title{Autoplot method for clipped linear fit}
\usage{
\method{autoplot}{clipped_lm_fit}(object, ...)
}
\arguments{
\item{object}{A \code{clipped_lm_fit} object created by \code{\link[=as.clipped_lm_fit]{as.clipped_lm_fit()}}.}

\item{...}{Additional arguments (unused).}
}
\value{
A \code{ggplot2} layer (a clipped \code{geom_segment}).
}
\description{
Autoplot method for clipped linear fit
}
