% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osmotic_estimates.R
\name{autoplot.osmEst}
\alias{autoplot.osmEst}
\title{Plot method of class osmEst}
\usage{
\method{autoplot}{osmEst}(
  object,
  xlab = "100-RWC",
  ylab = expression(Psi),
  main = "Hofler Diagram",
  lwd = 1.5,
  lty = 1,
  ...
)
}
\arguments{
\item{object}{Object of class osmEst}

\item{xlab}{x label}

\item{ylab}{y label}

\item{main}{Title main}

\item{lwd}{line width}

\item{lty}{line type}

\item{...}{unused}
}
\value{
Hofler diagram of osmotic elements
}
\description{
Plot method of class osmEst
}
