% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoplot.pv.R
\name{autoplot.pv}
\alias{autoplot.pv}
\title{Plotting method for pv objects}
\usage{
\method{autoplot}{pv}(
  object,
  xlab = "100-RWC",
  ylab = expression(Psi),
  main = "Hofler Diagram",
  lwd = 1.5,
  lty = 1,
  ...
)
}
\arguments{
\item{object}{A pv object}

\item{xlab}{Label for x-axis}

\item{ylab}{Label for y-axis}

\item{main}{Main title for the plot}

\item{lwd}{Line width}

\item{lty}{Line type}

\item{...}{Additional arguments passed to autoplot generic}
}
\value{
A ggplot object representing the Hofler diagram
}
\description{
Plotting method for pv objects
}
