% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estPV.R
\name{by_grp_sbgrp}
\alias{by_grp_sbgrp}
\title{Separate input dataframe by group and subgroup}
\usage{
by_grp_sbgrp(x, grp, sbgrp = NULL)
}
\arguments{
\item{x}{A dataframe}

\item{grp}{Grouping variable}

\item{sbgrp}{Subgrouping variable}
}
\description{
Separate input dataframe by group and subgroup
}
