% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sma_slope_intercept.R
\name{calc_param_tlp}
\alias{calc_param_tlp}
\title{Calculate the slopes for turgor loss inputs}
\usage{
calc_param_tlp(x, y = NULL, ...)
}
\arguments{
\item{x}{Object of class \emph{\code{tlp_input}}}
}
\value{
Calculated slopes and intercepts relating pressure potential with relative water deficit, and relative water content with water potential above and below turgor loss point. See attr(x, "model_list") for explicit display of fitted models and data used.
}
\description{
This function estimates the standard major axis parameters used in predicting values at water potential at turgor loss
}
\seealso{
\code{\link[=estTLP]{estTLP()}}
}
