% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_n_points.R
\name{check_n_pts}
\alias{check_n_pts}
\title{Check the number of rows that produces the highest R^2 values}
\usage{
check_n_pts(data, wp.index, wm.index, max_row = nrow(data))
}
\arguments{
\item{data}{a dataframe}

\item{wp.index}{index number or string containing the column name of the water potential data}

\item{wm.index}{index number or string containing the column name of the water mass data (i.e., water mass or relative water contentt)}

\item{max_row}{set cap on the number of rows to test for best r-squared value. Default to number of rows in the \code{data}}
}
\value{
returns a list of two with the number of rows that produces the greatest r-squared (r2) or lowest pi_o value
}
\description{
Check the number of rows that produces the highest R^2 values
}
\concept{internal}
