% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data}
\alias{data}
\title{Data}
\format{
Data frame with columns
\describe{
\item{species}{Species code: first two letters of Genus and Species}
\item{leaf}{Leaf number}
\item{dry.weight}{Dry mass of leaf in grams}
\item{water.potential}{:eaf water potential in MPa}
\item{fresh.weight}{Leaf water weight in grams}
}
}
\usage{
data
}
\description{
Pressure volume curve data for Quercus agrifolia
}
\examples{
head(data)
}
\keyword{datasets}
