% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggpv.R
\name{drawSegments}
\alias{drawSegments}
\title{Draw line segments on plots}
\usage{
drawSegments(fit, x_range, color = "black", ...)
}
\arguments{
\item{fit}{A fit object from a object of class PVest}

\item{x_range}{Range of independent variable}

\item{color}{Line color}

\item{...}{Additional parameters passed to geom_segment}
}
\value{
Segments for the segmented regression
}
\description{
Draw line segments on plots
}
