% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tlp_estimates.R
\name{EstimateTLP}
\alias{EstimateTLP}
\title{Estimate parameters at turgor loss}
\usage{
EstimateTLP(df, fw.index, wp.index, n_row = 4)
}
\arguments{
\item{df}{A data frame}

\item{fw.index}{Water associated variable}

\item{wp.index}{Pressure associated variable}
}
\value{
Returns data with new columns for the estimated parameters.
}
\description{
Estimate leaf water potential, relative water content at turgor loss point and modulus of elasticity
both bulk parameters and their symplastic counterparts
}
