% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osmotic_estimates.R
\name{estOsmotic}
\alias{estOsmotic}
\alias{estOsmotic.default}
\alias{estOsmotic.rwcEst}
\title{Osmotic and pressure parameter estimation}
\usage{
estOsmotic(x, ...)

\method{estOsmotic}{default}(x, wc.index, wp.index, n_row = 4, silent = T, ...)

\method{estOsmotic}{rwcEst}(x, n_row = 4, silent = T, ...)
}
\arguments{
\item{x}{An object of class "rwcEst"}

\item{...}{Additional parameters passed to method (unused)}

\item{wc.index}{A double. A data frame index which contains the water content (RWD) data passed to estpio()}

\item{wp.index}{A double. A data frame index which contains the leaf water potential data}

\item{n_row}{Number of rows above turgor loss}

\item{silent}{Silence printing of head(data) and columns used}
}
\value{
Returns data frame with new columns containing the osmotic and pressure potential variables namely:
\itemize{
\item osmotic potential at full turgor
\item maximum pressure potential
\item osmotic potential (for each hydration state)
\item pressure potential (for each hydration state)
\item apoplastic fraction
\item symplastic relative water content(sym.rwc; for each hydration state)
}
}
\description{
Estimate the osmotic potential at full turgor, pressure potential at full turgor for leaves
as well as the osmotic potential and pressure potential for each hydration state.
}
\details{
This function estimates osmotic variables from the values \strong{below} turgor loss point. Here we assume that the data points at
last 4 hydration states all represent points below turgor loss point.

It is implemented \strong{after} estimation of leaf saturated water content, relative water content and relative water deficit.See \code{estRWC} for information on the estimation of those parameters.

Here, we estimate the osmotic potential at full turgor as the x-intercept of the relationship between
inverse leaf water potential and relative water deficit below turgor loss point.
}
\seealso{
\code{\link[=estRWC]{estRWC()}}, \code{\link[=sma_model]{sma_model()}}
}
