% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estParams.R
\name{estParams}
\alias{estParams}
\title{Estimate all pressure volume curve parameters for all leaves}
\usage{
estParams(
  data,
  fw.index,
  wp.index,
  dm.index,
  method = "cv",
  max_row = nrow(data)
)
}
\arguments{
\item{data}{A data frame. Input data frame with the raw values}

\item{fw.index}{A numeric. Data frame index for the water content information.}

\item{wp.index}{A numeric. Indicates the index of the data frame including the water potential data}

\item{dm.index}{Numeric index of dry mass data.}

\item{method}{Default "cv"; See details for more information. If method="cv", the number of rows to use for predicting pio is based on the lowest number of rows with a CV less than 10\%.}

\item{max_row}{Defaults to length of the input data frame. Maximum number of rows to check with method.}
}
\value{
Returns a data frame with the estimated parameters
}
\description{
Estimate all pressure volume curve parameters for all leaves
}
\details{
\if{html}{\out{<div class="sourceCode">}}\preformatted{Data needs to have species or individual information as well as leaf identifiers.
The function takes your "water.potential" column(recommend naming it thus) and a "fresh water content"(again, ideally named fresh.water.content) for multiple
hydration states. For information on the production of leaf pressure volume curves see Sack 2011.
With the data for each leaf it takes the first for values and estimates a saturated water content, estimates
relative water content and then with the relationship between relative water deficit and 1/psi, it estimates the
osmotic potential and pressure potential at full turgor(max.psip) from which the same parameters are estimated for each
subsequent hydration state.

Method options: 
* "cv" - number of rows chosen by lowest number of points where slope of inv_psi vs RWD is greater than 10\%
* "r2" - number of rows chosen as the number of points where R^2 is greatest.
* "lowest" - number of rows chosen as number of points that estimates the lowest pio value between 4-max_rows rows. 
* "median" - number of rows chosen as number of points that estimates the median pio value between 4-max_rows rows. 
* if nothing specified - estimate pio with only 4 observations. 
}\if{html}{\out{</div>}}
}
