% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osmotic_estimates.R
\name{estpio}
\alias{estpio}
\title{Osmotic potential at full turgor estimate}
\usage{
estpio(rwd, psi)
}
\arguments{
\item{rwd}{Vector of relative water deficit values.}

\item{psi}{Vector of water potential values.}
}
\value{
An object of class "pioEst" containing the slope, intercept and osmotic potential at full turgor.
}
\description{
Estimates the osmotic potential at full turgor from the linear relationship
between relative water deficit and inverse leaf water potential.
}
\seealso{
\code{\link[=sma_model]{sma_model()}}
}
