% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estPV.R
\name{estPV}
\alias{estPV}
\alias{estPV.default}
\alias{print.estPV}
\title{Estimate all pressure volume curve parameters for all leaves}
\usage{
estPV(data, group, subgrp = NULL, fw, wp, dm, method = NULL)

\method{estPV}{default}(data, group, subgrp = NULL, fw, wp, dm, method = NULL)

\method{print}{estPV}(x, ...)
}
\arguments{
\item{data}{A data frame. Input data frame with the raw values}

\item{group}{An unquoted character. The column name of the grouping variable}

\item{subgrp}{An unquoted character. The column name of the subgrouping variable}

\item{fw}{A numeric. Data frame index for the water content information.}

\item{wp}{A numeric. Indicates the index of the data frame including the water potential data}

\item{dm}{Numeric index of dry mass data.}

\item{method}{Optional character parameter for when n_pts=T. Determines which method to pick number of rows below TLP.Options include "r2","pio","cv". See \code{?check_n_points} for more information}

\item{x}{Object of class \emph{estPV}}

\item{...}{Additional parameters passed to method}

\item{n_pts}{Logical. If \code{TRUE}, the number of rows to use for predicting pio is based on the lowest number of rows with a CV less than 10\%.}
}
\value{
Returns a data frame with the estimated parameters
}
\description{
Estimate all pressure volume curve parameters for all leaves
}
\details{
\if{html}{\out{<div class="sourceCode">}}\preformatted{Data needs to have species or individual information as well as leaf identifiers.
The function takes your "water.potential" column(recommend naming it thus) and a "fresh water content"(again, ideally named fresh.water.content) for multiple
hydration states. For information on the production of leaf pressure volume curves see Sack 2011.
With the data for each leaf it takes the first for values and estimates a saturated water content, estimates
relative water content and then with the relationship between relative water deficit and -1/psi, it estimates the
osmotic potential and pressure potential at full turgor(max.psip) from which the same parameters are estimated for each
subsequent hydration state.
}\if{html}{\out{</div>}}
}
