% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saturated_water_content.R
\name{estRWC}
\alias{estRWC}
\title{RWC and RWD estimation}
\usage{
estRWC(data, fw.index, wp.index, dm.index, n_row = 4, silent)
}
\arguments{
\item{data}{Data frame containing leaf fresh water mass and leaf water potentials}

\item{fw.index}{Numeric value indicating the column number where the leaf water mass data is within data frame}

\item{wp.index}{Water potential column index}

\item{dm.index}{Dry mass column index}

\item{silent}{Silence printing of column names}
}
\value{
Returns a list of two containing the relative water content(RWC) for each measurement point and
the relative water deficit(i.e, 100-RWC)
}
\description{
Estimate the relative water content(RWC) and relative water deficit(RWD) based on the
saturated water content. For a
detailed explanation of the function see \code{vignette('swc-and-rwc-estimation')}
}
\seealso{
Other abovetlp: 
\code{\link{estsatwater}()}
}
\concept{abovetlp}
