% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saturated_water_content.R
\name{estsatwater}
\alias{estsatwater}
\title{Saturated water content(SWC) estimation}
\usage{
estsatwater(fresh_mass, psi, dry_mass)
}
\arguments{
\item{fresh_mass}{Numeric vector of fresh water mass}

\item{psi}{Numeric vector of water potential}

\item{dry_mass}{Leaf dry mass}
}
\value{
Returns the saturated water content from the relationship between water mass and water potential
}
\description{
Estimate the saturated water content from the intercept of the
the relationship between leaf water mass and leaf water potential. For a
detailed explanation of the function see \code{vignette('swc-and-rwc-estimation')}
}
\details{
This function calculates the saturated water mass (swm) and saturated water content (swc) as the standard major axis intercept.

The saturated water mass is equivalent to the y-intercept of the standard major axis line relating water potential (\eqn{\Psi}) to leaf water content(lwc), or :

\deqn{swm = \beta_{SMA}}

The saturated water content (mass as a function of total leaf water) is then calculated as:

\deqn{swc = \frac{swm}{dry~mass}}
}
\seealso{
\itemize{
\item \code{\link[=sma_intercept]{sma_intercept()}} Function used for estimating the SMA intercept
\item \code{\link[=sma_slope]{sma_slope()}} Function used for estimating the SMA slope
\item \code{\link[=estRWC]{estRWC()}} Estimates the relative water content and relative water deficit using the saturated water content estimate here
}
}
