% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tlp_estimates.R
\name{estTLP.default}
\alias{estTLP.default}
\title{Default method for estTLP}
\usage{
\method{estTLP}{default}(data, wc.index, wp.index, n_row_above = 4, n_row_below = 4)
}
\description{
Default method for estTLP
}
