% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tlp_estimates.R
\name{estTLP.osmEst}
\alias{estTLP.osmEst}
\title{estTLP method}
\usage{
\method{estTLP}{osmEst}(osm_object, wc.index, wp.index, n_row_above = 4, n_row_below = 4)
}
\description{
A method of estTLP for use with objects that are of class "osmEst"(i.e., returned from pvest::estOsm())
}
