% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osmotic_estimates.R
\name{get_varnames}
\alias{get_varnames}
\title{Get varnames}
\usage{
get_varnames(data, wc.index, wp.index)
}
\arguments{
\item{data}{Input data frame}

\item{wc.index}{Selected water content index}

\item{wp.index}{Selected water potential index}
}
\value{
Variable names
}
\description{
Get varnames
}
