% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataCheck.R
\name{InputCheck}
\alias{InputCheck}
\title{Check the condition of the data frame}
\usage{
InputCheck(data)
}
\arguments{
\item{data}{A data frame}
}
\value{
Stops evaluation of function
}
\description{
Check the condition of the data frame
}
