% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osmotic_estimates.R
\name{new_osminput}
\alias{new_osminput}
\title{Constructor for osmotic inputs}
\usage{
new_osminput(rwd, psi)
}
\arguments{
\item{rwd}{Values of relative water content and deficit}

\item{psi}{Values of water potential}
}
\description{
Constructor for osmotic inputs
}
