% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optim_thres.R
\name{optim_thres}
\alias{optim_thres}
\title{Determine best number of points below turgor loss}
\usage{
optim_thres(data, fw, wp, dm, method = "rmse")
}
\arguments{
\item{data}{a data frame}

\item{method}{Specify statistic used to optimize the number of rows above and below turgor loss.}

\item{x_var}{number or string containing the column name of the water potential data}

\item{y_var}{index number or string containing the column name of the water mass data (i.e., water mass or relative water content)}
}
\value{
Returns list with:
\itemize{
\item the objective cost function value for each index that was used to minimize (i.e., rmse or aicc) or maximized (R squared). The attributes of which are the number of rows on either side of turgor used to fit the standard major axis model.
\item Number of rows above (\code{n_above}) and below (\code{n_below}) that optimized the objective function.
}
}
\description{
This function aims to determine the best number of points by coefficients of determination, coefficients of variation, and estimated leaf osmotic potential at full turgor.
}
\concept{internal}
