% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osmotic_estimates.R
\name{OsmoticPotFullTurgor}
\alias{OsmoticPotFullTurgor}
\title{Osmotic potential at full turgor estimate}
\usage{
OsmoticPotFullTurgor(data, fw.index, wp.index)
}
\arguments{
\item{data}{}

\item{fw.index}{}

\item{wp.index}{}
}
\value{
Numeric list containing the slope, intercept and osmotic potential at full turgor, in that order.
}
\description{
Estimates the osmotic potential at full turgor from the linear relationship
between relative water deficit and inverse leaf water potential.
}
