% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piecewise_regression.R
\name{piecewise_reg}
\alias{piecewise_reg}
\title{Piecewise linear regression estimate}
\usage{
piecewise_reg(df)
}
\arguments{
\item{df}{A data.frame containing inverse water potential and relative water deficit}
}
\value{
Returns list of two best fitting models over the two sections of data
}
\description{
Iterate through possible segment divisions
}
