% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggpv.R
\name{plotPV}
\alias{plotPV}
\title{Stat function for plotting pv curve on data}
\usage{
plotPV(obj, ...)
}
\arguments{
\item{obj}{object of class estPV}

\item{...}{additional arguments not used}
}
\value{
A list of patchwork plots (length = number of species) of water potential and water content relationships
}
\description{
Stat function for plotting pv curve on data
}
