% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osmotic_estimates.R
\name{print.pioEst}
\alias{print.pioEst}
\title{Print method for the Osmotic estimates output}
\usage{
\method{print}{pioEst}(x, ...)
}
\arguments{
\item{x}{An object of the class "pioEst"}

\item{...}{Other parameters passed to cat}
}
\value{
Printed SMA model output
}
\description{
Print method for the Osmotic estimates output
}
