% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sma_slope_intercept.R
\name{print.sma_model}
\alias{print.sma_model}
\title{Print method for the SMA model output}
\usage{
\method{print}{sma_model}(x, ...)
}
\arguments{
\item{x}{An object of the class "sma_model"}

\item{...}{Other parameters passed to cat}
}
\value{
Printed SMA model output
}
\description{
Print method for the SMA model output
}
