% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tlp_estimates.R
\name{psip_rwd_params}
\alias{psip_rwd_params}
\title{Parameters to estimate values at turgor loss point from leaf pressure potential versus bulk and symplastic relative water deficit SMA parameters.}
\usage{
psip_rwd_params(psi, psip, rwd, sym_rwd, rwc, sym_rwc)
}
\arguments{
\item{psip}{A vector of pressure potential values}

\item{rwd}{A vector of relative water deficit values}

\item{symrwd}{A vector of symplastic relative water deficit values}
}
\value{
Bulk and symplastic slopes and intercepts.
}
\description{
Parameters to estimate values at turgor loss point from leaf pressure potential versus bulk and symplastic relative water deficit SMA parameters.
}
