% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tlp_estimates.R
\name{PsiPRWD_slopeint}
\alias{PsiPRWD_slopeint}
\title{Estimate the RWC and RWD at turgor loss point for using the relationship between leaf pressure potential and
bulk and symplastic relative water deficit}
\usage{
PsiPRWD_slopeint(psip.index, rwd.index, sym_rwd.index)
}
\arguments{
\item{psip.index}{TBD}

\item{rwd.index}{TBD}

\item{sym_rwd.index}{TBD}
}
\value{
Bulk and symplastic slopes and intercepts
}
\description{
Estimate the RWC and RWD at turgor loss point for using the relationship between leaf pressure potential and
bulk and symplastic relative water deficit
}
