% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{quag}
\alias{quag}
\title{quag
Pressure volume curve data for Quercus agrifolia}
\format{
Data frame with columns
\describe{
\item{species}{Species code: first two letters of Genus and Species}
\item{leaf}{Leaf number}
\item{dry.weight}{Dry mass of leaf in grams}
\item{water.potential}{:eaf water potential in MPa}
\item{fresh.weight}{Leaf water weight in grams}
}
}
\usage{
quag
}
\description{
quag
Pressure volume curve data for Quercus agrifolia
}
\keyword{datasets}
