% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saturated_water_content.R
\name{SaturatedWaterContent}
\alias{SaturatedWaterContent}
\title{SWC estimations}
\usage{
SaturatedWaterContent(data, fw.index, wp.index)
}
\arguments{
\item{data}{Data frame containing leaf fresh water mass and leaf water potentials}

\item{fw.index}{Numeric value indicating the column number where the leaf water mass data is within data frame}

\item{wp.index}{Numeric value indicating the column number where the leaf water potential data is within the data frame}
}
\value{
Returns the saturated water content from the relationship between water mass and water potential
}
\description{
Estimate the saturated water content from the intercept of the
the relationship between leaf water mass and leaf water potential.
}
\details{
This function calculates the saturated water content as the standard major axis intercept.
}
\seealso{
\itemize{
\item \link{sma_intercept} Function used for estimating the SMA intercept
\item \link{sma_slope} Function used for estimating the SMA slope
\item \link{RelativeWaterCD} Estimates the relative water content and relative water deficit using the saturated water content estimate here
}

Other abovetlp: 
\code{\link{RelativeWaterCD}()}
}
\concept{abovetlp}
