% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sma_slope_intercept.R
\name{sma_intercept}
\alias{sma_intercept}
\title{SMA intercept estimation}
\usage{
sma_intercept(x, y, slope)
}
\arguments{
\item{x}{x variable}

\item{y}{y variable}

\item{slope}{slope parameter estimated by the 'sma_slope' function}
}
\value{
Returns slope and intercept values for standard major axis regression
}
\description{
SMA intercept estimation
}
