% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sma_slope_intercept.R
\name{sma_model.default}
\alias{sma_model.default}
\title{Default SMA model formulation}
\usage{
\method{sma_model}{default}(x, y, ...)
}
\arguments{
\item{x}{independent variable}

\item{y}{dependent variable}
}
\value{
Returns a model function slope and intercept values for the standard major axis regression
}
\description{
Default SMA model formulation
}
