% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sma_slope_intercept.R
\name{sma_model.osm_input}
\alias{sma_model.osm_input}
\title{Method for calculate SMA model parameters for determining osmotic parameters}
\usage{
\method{sma_model}{osm_input}(x, y)
}
\arguments{
\item{x}{independent variable}

\item{y}{dependent variable}
}
\value{
Returns a model function slope and intercept values for the standard major axis regression
}
\description{
Method for calculate SMA model parameters for determining osmotic parameters
}
