% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sma_slope_intercept.R
\name{sma_model}
\alias{sma_model}
\title{SMA model formulation}
\usage{
sma_model(x, y)
}
\arguments{
\item{x}{independent variable}

\item{y}{dependent variable}
}
\value{
Returns a model function slope and intercept values for the standard major axis regression
}
\description{
SMA model formulation
}
