% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sma_slope_intercept.R
\name{sma_model2}
\alias{sma_model2}
\title{Two Standard Major Axis Fits}
\usage{
sma_model2(dfs, wp, lw, method = "default", pmetric = NULL)
}
\arguments{
\item{dfs}{List of dataframe for above and below turgor loss}

\item{wp}{String indicating water potential index}

\item{lw}{String indicating leaf water content measurement}

\item{method}{Method of determining where to split data}

\item{pmetric}{Significance metric usually pvalues}
}
\description{
Two Standard Major Axis Fits
}
\details{
See \code{?sma_model} for more information on how the SMA models are fit.
}
