% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sma_slope_intercept.R
\name{sma_slope}
\alias{sma_slope}
\title{SMA slope}
\usage{
sma_slope(x, y)
}
\arguments{
\item{x}{x variable}

\item{y}{y variable}
}
\value{
Returns slope value for standard major axis regression
}
\description{
Standard major axis(SMA) parameter estimation. The functions provided here
will estimate the slope and intercept of the SMA line for a given dataset. For a
detailed explanation of the function see \code{vignette('sma-slope-intercept')}
}
\details{
The slope of the relationship is determined as the the standard deviation of
of the x variable divided by the standard deviation of the y variable. Estimation of the intercept
requires the estimation of the SMA slope. The intercept is estimated as the mean of the x variable
minus the SMA slope times the mean value of the y variable. Often in pressure volume curves, the x variable is the water-associated
variable(e.g., water content or leaf relative water content). Likewise, the y variable is often associated
with pressure(e.g., leaf water potential, and inverse of leaf water potential)
}
