% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitdf.R
\name{split_all_dfs}
\alias{split_all_dfs}
\title{Split Data Frame by Vector}
\usage{
split_all_dfs(df_list, split_vectors)
}
\arguments{
\item{df_list}{A list of data frames to be split.}

\item{split_vectors}{A list of vectors indicating the number of rows to take from the head and tail of each data frame.}
}
\value{
A list of lists, where each inner list contains the split data frames.
}
\description{
Split Data Frame by Vector
}
