% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitdf.R
\name{split_df}
\alias{split_df}
\title{Split Data Frame by Vector}
\usage{
split_df(df, split_vector)
}
\arguments{
\item{df}{A data frame to be split.}

\item{split_vector}{A vector of two integers indicating the number of rows to take from the head and tail of the data frame.}
}
\value{
A list containing two data frames: one with the specified number of rows from the head and another with the specified number of rows from the tail.
}
\description{
Split Data Frame by Vector
}
