% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggpv.R
\name{stat_two_part_line}
\alias{stat_two_part_line}
\title{Plot a two-part line based on extracted model elements}
\usage{
stat_two_part_line(
  mapping = NULL,
  data = NULL,
  geom = "path",
  position = "identity",
  model_fun = NULL,
  attr_name = "line_params",
  ...,
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE
)
}
\arguments{
\item{mapping}{Set of aesthetic mappings}

\item{data}{The data to be displayed in this layer}

\item{geom}{The geometric object to use display the data}

\item{position}{Position adjustment}

\item{model_fun}{Function that returns model with slope/intercept attributes}

\item{attr_name}{The name of the attribute containing slope/intercept values}

\item{...}{Other arguments passed on to \code{layer()}}

\item{na.rm}{If \code{FALSE}, the default, missing values are removed with
a warning. If \code{TRUE}, missing values are silently removed.}

\item{show.legend}{logical. Should this layer be included in the legends?}

\item{inherit.aes}{If \code{FALSE}, overrides the default aesthetics}
}
\description{
This stat extracts slope and intercept values from attributes of a model object
and creates a visualization of a two-part line.
}
