% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggpv.R
\docType{data}
\name{statPV}
\alias{statPV}
\title{Stat function for plotting pv curve on data}
\format{
An object of class \code{statPV} (inherits from \code{Stat}, \code{ggproto}, \code{gg}) of length 2.
}
\usage{
statPV
}
\description{
Stat function for plotting pv curve on data
}
\keyword{datasets}
