% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osmotic_estimates.R
\name{summary.osmEst}
\alias{summary.osmEst}
\title{Print method of osmEst class objects}
\usage{
\method{summary}{osmEst}(object, ...)
}
\arguments{
\item{object}{Object of class osmEst}
}
\description{
Print method of osmEst class objects
}
