% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osmotic_estimates.R
\name{summary.pioEst}
\alias{summary.pioEst}
\title{Summary method for pioEst class objects}
\usage{
\method{summary}{pioEst}(object, ...)
}
\arguments{
\item{object}{An object of class "pioEst"}

\item{...}{Other parameters passed to summary()}
}
\value{
The underlying SMA model object
}
\description{
Summary method for pioEst class objects
}
