% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sma_slope_intercept.R
\name{summary.sma_model}
\alias{summary.sma_model}
\title{Print method for the SMA model output}
\usage{
\method{summary}{sma_model}(object, ...)
}
\arguments{
\item{object}{Object of class "sma_model"}

\item{...}{Other parameters passed to print method}
}
\value{
Printed SMA model output
}
\description{
Print method for the SMA model output
}
