% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sma_slope_intercept.R
\name{summary.sma_model2}
\alias{summary.sma_model2}
\title{Print method for SMA model with two segments}
\usage{
\method{summary}{sma_model2}(object, ...)
}
\arguments{
\item{object}{An object of class \code{sma_model2}}

\item{...}{Additional arguments (not used)}
}
\value{
Prints the model summary
}
\description{
Print method for SMA model with two segments
}
