% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osmotic_estimates.R
\name{calc_nonlin_psip}
\alias{calc_nonlin_psip}
\title{Nonlinear pressure potential estimate}
\usage{
calc_nonlin_psip(data, pi_sat, r_tlp, psi_w = NULL, full = FALSE)
}
\arguments{
\item{data}{A data frame. A data frame containing the data set of the hydration states for a given leaf above turgor loss estimate}

\item{pi_sat}{A double. Osmotic potential at full turgor}

\item{r_tlp}{A double. Relative water content at turgor loss point}

\item{psi_w}{A double. Leaf water potential (optional, used if full = TRUE)}

\item{full}{Logical. If TRUE, fits the full model including water potential, otherwise fits only the pressure potential model}
}
\value{
Pressure potential estimator
}
\description{
Estimate the pressure potential using a nonlinear model
}
\details{
Here we fit a nonlinear model to estimate the pressure potential based on the osmotic potential at full turgor where the pressure potential decreases nonlinearly before turgor loss with a slope of b :
$$\Psi_p = -\pi_{sat} \left(\frac{RWC - RWC_{TLP}}{1 - RWC_{TLP}}\right)^b$$
}
