% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/af_sym_cap.R
\name{capacitance_fttlp}
\alias{capacitance_fttlp}
\title{Estimate the capacitance at full turgor and turgor loss point}
\usage{
capacitance_fttlp(data, ft_tlp)
}
\arguments{
\item{data}{A data frame}

\item{ft_tlp}{A string. "ft" indicates capacitance at full turgor and "tlp"
indicates leaf capacitance at turgor loss point}
}
\value{
Returns data frame
}
\description{
Estimate the capacitance at full turgor and turgor loss point
}
