% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_n_points.R
\name{check_n_pts}
\alias{check_n_pts}
\title{Determine best number of points below turgor loss}
\usage{
check_n_pts(data, wp.index, wm.index, max_row = 10)
}
\arguments{
\item{data}{a data frame}

\item{wp.index}{index number or string containing the column name of the water potential data}

\item{wm.index}{index number or string containing the column name of the water mass data (i.e., water mass or relative water content)}

\item{max_row}{set cap on the number of rows to test for best r-squared value. Default to 10 rows of data.}
}
\value{
returns a list of 6 containing r2 values, osmotic potentials and coefficients of variation. See below for further information.

\bold{Return values}

\itemize{
\item all_r2. Coefficient of determination (R^2^) for each number of rows
\item cv. Coefficient of variation (CV) in the slope between increasing number of rows
\item r2. N rows where there is the greatest R^2^
\item pi_o. N rows that predicts the most negative osmotic potential at full turgor.
\item cv10. Rows where the CV in slope in less than 10\%
}
}
\description{
This function aims to determine the best number of points by coefficients of determination, coefficients of variation, and estimated leaf osmotic potential at full turgor.
}
\details{
An alternative to the default number of rows would be to set \code{max_rows} to number of rows in the \code{data}. Note that more rows require greater computational time especially if computing parameters across many species. Suggest running the function with the default then, if a greater number of rows need to be tested, rerun the function for a subset of species.
}
\concept{internal}
