% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tlp_estimates.R
\name{EstimateTLP}
\alias{EstimateTLP}
\title{Estimate parameters at turgor loss}
\usage{
EstimateTLP(df, wc.index, wp.index, n_row = 4)
}
\arguments{
\item{df}{A data frame}

\item{wc.index}{An unquoted string indicating vector with water associated variable}

\item{wp.index}{An unquoted string indicating vector with pressure associated variable}

\item{n_row}{A double. Default to n_row=4 (i.e., 4 rows).Value which indicates the number of rows for estimating SMA line parameters.}
}
\value{
Returns data with new columns for the estimated parameters.
}
\description{
Estimate leaf water potential, relative water content at turgor loss point and modulus of elasticity
both bulk parameters and their symplastic counterparts
}
