% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osmotic_estimates.R
\name{estOsmotic}
\alias{estOsmotic}
\title{Osmotic and pressure parameter estimation}
\usage{
estOsmotic(
  data,
  wc.index = "relative.water.deficit",
  wp.index = "inv.water.potential",
  n_row = 4
)
}
\arguments{
\item{data}{A data frame. A data frame containing the data set of the last 4 hydration states for a given leaf}

\item{wc.index}{A double. A data frame index which contains the water content data passed through OsmoticPotFullTurgor function}

\item{wp.index}{A double. A data frame index which contains the leaf water potential data}

\item{n_row}{A double. Value which indicates the number of rows for estimating parameters. Default to 4 rows.}
}
\value{
Returns data frame with new columns containing the osmotic and pressure potential variables namely:
\itemize{
\item osmotic potential at full turgor
\item maximum pressure potential
\item osmotic potential (for each hydration state)
\item pressure potential (for each hydration state)
\item apoplastic fraction
\item symplastic relative water content(sym.rwc; for each hydration state)
}
}
\description{
Estimate the osmotic potential at full turgor, pressure potential at full turgor for leaves
as well as the osmotic potential and pressure potential for each hydration state.
}
\details{
This function estimates osmotic variables from the values \strong{below} turgor loss point. Here we assume that the data points at
last 4 hydration states all represent points below turgor loss point.

It is implemented \strong{after} estimation of leaf saturated water content, relative water content and relative
water deficit.See \code{RelativeWaterCD} for information on the estimation of those parameters.

Here, we estimate the osmotic potential at full turgor as the x-intercept of the relationship between
inverse leaf water potential and relative water deficit below turgor loss point. The slope needs to be negative
within the sma_intercept function(\code{\link{sma_intercept}}) because of the negative inverse of leaf water potential.
Also, note that unlike in the saturated water content estimation the x values is inverse.
}
\seealso{
\code{\link[=RelativeWaterCD]{RelativeWaterCD()}}, \code{\link[=sma_intercept]{sma_intercept()}}
}
