% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estParams.R
\name{estParams}
\alias{estParams}
\title{Estimate all pressure volume curve parameters for all leaves}
\usage{
estParams(data, fw.index, wp.index)
}
\arguments{
\item{data}{A data frame. Input data frame with the raw values}

\item{fw.index}{A numeric. Data frame index for the water content information.}

\item{wp.index}{A numeric. Indicates the index of the data frame including the water potential data}
}
\value{
Returns a data frame with the estimated parameters
}
\description{
Estimate all pressure volume curve parameters for all leaves
}
\details{
\if{html}{\out{<div class="sourceCode">}}\preformatted{Data
}\if{html}{\out{</div>}}
}
