% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osmotic_estimates.R
\name{estpio}
\alias{estpio}
\title{Osmotic potential at full turgor estimate}
\usage{
estpio(rwd, psi)
}
\arguments{
\item{rwd}{Vector of relative water deficit values.}

\item{psi}{Vector of water potential values.}
}
\value{
Numeric list containing the slope, intercept and osmotic potential at full turgor, in that order.Output is an object of class "pioEst"
}
\description{
Estimates the osmotic potential at full turgor from the linear relationship
between relative water deficit and inverse leaf water potential.
}
\seealso{
\code{\link{sma_slope}}, \code{\link{sma_intercept}}
}
