% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osmotic_estimates.R
\name{estpio}
\alias{estpio}
\title{Osmotic potential at full turgor estimate}
\usage{
estpio(rwc, psi)
}
\arguments{
\item{rwc}{A vector of relative water content or deficit data}

\item{psi}{A vector of water potential data}

\item{data}{A data frame. A data frame containing the data set of the last 4 hydration states for a given leaf}
}
\value{
Numeric list containing the slope, intercept and osmotic potential at full turgor, in that order. Class "osmEst"
}
\description{
Estimates the osmotic potential at full turgor from the linear relationship
between relative water deficit and inverse leaf water potential.
}
\seealso{
\code{\link{sma_slope}}, \code{\link{sma_intercept}}
}
