% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tlp_estimates.R
\name{estTLP}
\alias{estTLP}
\alias{estTLP.default}
\alias{estTLP.osmEst}
\title{Estimate parameters at turgor loss}
\usage{
estTLP(x, ...)

\method{estTLP}{default}(data, wc.index, wp.index, n_row_above = 4, n_row_below = 4)

\method{estTLP}{osmEst}(osm_object, wc.index, wp.index, n_row_above = 4, n_row_below = 4)
}
\arguments{
\item{data}{A data frame}

\item{wc.index}{An unquoted string indicating vector with water associated variable}

\item{wp.index}{An unquoted string indicating vector with pressure associated variable}

\item{n_row_above}{Double; Default to 4 rows.Value which indicates the number of rows above TLP for estimating SMA line parameters.}

\item{n_row_below}{Double; Default to 4 rows.Value which indicates the number of rows below TLP for estimating SMA line parameters.}
}
\value{
Returns data with new columns for the estimated parameters.
}
\description{
Estimate leaf water potential, relative water content at turgor loss point and modulus of elasticity both bulk parameters and their symplastic counterparts
}
