% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osmotic_estimates.R
\name{osminput}
\alias{osminput}
\title{Constructor for osmotic inputs}
\usage{
osminput(rwd, psi)
}
\arguments{
\item{psi}{Values of water potential}

\item{rwc}{Values of relative water content and deficit}
}
\description{
Constructor for osmotic inputs
}
