% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osmotic_estimates.R
\name{OsmoticPotFullTurgor}
\alias{OsmoticPotFullTurgor}
\title{Osmotic potential at full turgor estimate}
\usage{
OsmoticPotFullTurgor(data, fw.index, wp.index)
}
\arguments{
\item{data}{A data frame. A data frame containing the data set of the last 4 hydration states for a given leaf}

\item{wp.index}{A double. Data frame index which contains the leaf water potential data}

\item{wc.index}{A double. Data frame index which contains the relative water content/deficit}
}
\value{
Numeric list containing the slope, intercept and osmotic potential at full turgor, in that order.
}
\description{
Estimates the osmotic potential at full turgor from the linear relationship
between relative water deficit and inverse leaf water potential.
}
\seealso{
\code{\link{sma_slope}}, \code{\link{sma_intercept}}
}
