% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPV.R
\name{plotPV}
\alias{plotPV}
\title{plotPV}
\usage{
plotPV(data, x, y, ...)
}
\arguments{
\item{data}{}

\item{x}{}

\item{y}{}

\item{...}{}
}
\value{
a plot
}
\description{
plotPV
}
\examples{

plotPV(data\%>\%filter(leaf=="5"), x='fresh.weight', y='water.potential')

}
