% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPV.R
\name{plotPV}
\alias{plotPV}
\title{plotPV}
\usage{
plotPV(data, x, y, rows = 4, ...)
}
\arguments{
\item{data}{dataframe}

\item{x}{x value}

\item{y}{y value}

\item{rows}{number of rows to select}

\item{...}{parameters to pass through plot()}
}
\value{
a plot
}
\description{
plotPV
}
