% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osmotic_estimates.R
\name{print.pioEst}
\alias{print.pioEst}
\title{Constructor for osmotic inputs}
\usage{
\method{print}{pioEst}(x, ...)
}
\arguments{
\item{x}{An object of the class "pioEst"}

\item{...}{Other parameters passed to cat}

\item{rwc}{Values of relative water content and deficit}

\item{psi}{Values of water potential

Print method for the Osmotic estimates output}
}
\value{
Printed SMA model output
}
\description{
Constructor for osmotic inputs
}
