% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tlp_estimates.R
\name{psip_rwd_params}
\alias{psip_rwd_params}
\title{Parameters to estimate values at turgor loss point from leaf pressure potential versus bulk and symplastic relative water deficit SMA parameters.}
\usage{
psip_rwd_params(
  psi_above,
  psip,
  rwd,
  symrwd,
  rwc_above,
  symrwc_above,
  psi_below,
  rwc_below,
  symrwc_below
)
}
\arguments{
\item{psi_above}{Water potential \emph{above} turgor loss}

\item{psip}{Pressure potential}

\item{rwd}{Relative water deficit}

\item{symrwd}{A vector of symplastic relative water deficit values}

\item{rwc_above}{Relative water content \emph{above} turgor loss}

\item{symrwc_above}{Symplastic relative water content \emph{above} turgor loss}

\item{psi_below}{Water potential \emph{below} turgor loss}

\item{rwc_below}{Relative water content \emph{below} turgor loss}

\item{symrwc_below}{Symplastic relative water content \emph{below} turgor loss}
}
\value{
Bulk and symplastic slopes and intercepts.
}
\description{
Parameters to estimate values at turgor loss point from leaf pressure potential versus bulk and symplastic relative water deficit SMA parameters.
}
