% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tlp_estimates.R
\name{psip_rwd_params}
\alias{psip_rwd_params}
\title{Estimate the RWC and RWD at turgor loss point for using the relationship between leaf pressure potential and bulk and symplastic relative water deficit}
\usage{
psip_rwd_params(psip, rwd, symrwd)
}
\arguments{
\item{psip}{A vector of pressure potential values}

\item{rwd}{A vector of relative water deficit values}

\item{symrwd}{A vector of symplastic relative water deficit values}
}
\value{
Bulk and symplastic slopes and intercepts.
}
\description{
Estimate the RWC and RWD at turgor loss point for using the relationship between leaf pressure potential and bulk and symplastic relative water deficit
}
