% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tlp_estimates.R
\name{PsiPRWD_slopeint}
\alias{PsiPRWD_slopeint}
\title{Estimate the RWC and RWD at turgor loss point for using the relationship between leaf pressure potential and
bulk and symplastic relative water deficit}
\usage{
PsiPRWD_slopeint(data, psip.index, rwd.index, sym_rwd.index)
}
\arguments{
\item{data}{A data frame. Must contain pressure potential and relative water deficit values.}

\item{psip.index}{A string or number.An indexing value for the vector in data for pressure potential.}

\item{rwd.index}{A string or number.An indexing value for the vector in data for relative water deficit.}

\item{sym_rwd.index}{A string or number.An indexing value for the vector in data for symplastic relative water deficit.}
}
\value{
Bulk and symplastic slopes and intercepts.
}
\description{
Estimate the RWC and RWD at turgor loss point for using the relationship between leaf pressure potential and
bulk and symplastic relative water deficit
}
