% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saturated_water_content.R
\name{RelativeWaterCD}
\alias{RelativeWaterCD}
\title{RWC and RWD estimation}
\usage{
RelativeWaterCD(data, fw.index, wp.index)
}
\arguments{
\item{data}{Data frame containing leaf fresh water mass and leaf water potentials}

\item{fw.index}{Numeric value indicating the column number where the leaf water mass data is within data frame}

\item{wp.index}{Numeric value indicating the column number where the leaf water potential data is within the data frame}
}
\value{
Returns the input dataframe with two new columns containing the relative water content(RWC) for each measurement point and
the relative water deficit(i.e, 100-RWC)
}
\description{
Estimate the relative water content(RWC) and relative water deficit(RWD) based on the
saturated water content
}
\seealso{
Other abovetlp: 
\code{\link{SaturatedWaterContent}()}
}
\concept{abovetlp}
