% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saturated_water_content.R
\name{SaturatedWaterContent}
\alias{SaturatedWaterContent}
\title{Saturated water content(SWC) estimation}
\usage{
SaturatedWaterContent(data, fw.index, wp.index, dm.index, n_row = 4)
}
\arguments{
\item{data}{Data frame containing leaf fresh water mass and leaf water potentials}

\item{fw.index}{Numeric value indicating the column number where the leaf water mass data is within data frame}

\item{wp.index}{Numeric value indicating the column number where the leaf water potential data is within the data frame}

\item{dm.index}{Numeric value indicating the column number where the leaf dry mass is}

\item{n_row}{A number. Indicates number of rows to estimate parameters from within data}
}
\value{
Returns the saturated water content from the relationship between water mass and water potential
}
\description{
Estimate the saturated water content from the intercept of the
the relationship between leaf water mass and leaf water potential. For a
detailed explanation of the function see \code{vignette('swc-and-rwc-estimation')}
}
\details{
This function calculates the saturated water content as the standard major axis intercept.
}
\seealso{
\itemize{
\item \code{\link[=sma_intercept]{sma_intercept()}} Function used for estimating the SMA intercept
\item \code{\link[=sma_slope]{sma_slope()}} Function used for estimating the SMA slope
\item \code{\link[=RelativeWaterCD]{RelativeWaterCD()}} Estimates the relative water content and relative water deficit using the saturated water content estimate here
}

Other abovetlp: 
\code{\link{RelativeWaterCD}()}
}
\concept{abovetlp}
