% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sma_slope_intercept.R
\name{sma_intercept}
\alias{sma_intercept}
\title{SMA intercept estimation}
\usage{
sma_intercept(x, y, slope)
}
\arguments{
\item{x}{x variable}

\item{y}{y variable}

\item{slope}{slope parameter estimated by the sma_slope function}
}
\value{
Returns slope and intercept values for standard major axis regression
}
\description{
SMA intercept estimation
}
\details{
Estimation of the intercept \strong{requires} the estimation of the SMA slope. The intercept is estimated as the mean of the x variable
minus the SMA slope times the mean value of the y variable. Often in pressure volume curves, the x variable is the water-associated
variable(e.g., water content or leaf relative water content). Likewise, the y variable is often associated
with pressure(e.g., leaf water potential, and inverse of leaf water potential). For a
detailed explanation of the function see \code{vignette('sma-slope-intercept')}
}
