% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sma_slope_intercept.R
\name{sma_model}
\alias{sma_model}
\title{SMA model generic}
\usage{
sma_model(x, y, ...)
}
\value{
Returns a model function slope and intercept values for the standard major axis regression
}
\description{
SMA model generic
}
