% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sma_slope_intercept.R
\name{sma_model}
\alias{sma_model}
\alias{sma_model.default}
\alias{sma_model.osm_input}
\alias{sma_model.tlp_input}
\title{Standard major axis model estimation.}
\usage{
sma_model(x, y, ...)

\method{sma_model}{default}(x, y, ...)

\method{sma_model}{osm_input}(x, y = NULL, ...)

\method{sma_model}{tlp_input}(x, y = NULL, ...)
}
\arguments{
\item{x}{independent variable}

\item{y}{dependent variable}

\item{...}{additional parameters passed to methods}
}
\value{
Returns a model function slope and intercept values for the standard major axis regression
}
\description{
This is a generic function that will dispatch 1 of 3 methods based on the input data type (i.e., \emph{default}, \emph{\code{osm_input}}, and \emph{\code{tlp_input}}) to construct a standard major axis model from calculated slope and intercept for a given dependent and independent variable:
\itemize{
\item \emph{default} : slope and intercept are positive
\item \emph{\code{osm_input}}: slope is forced negative representing the relationship between relative water deficit and the negative inverse of water potential. Intercept is positive.
\item \emph{\code{tlp_input}}: same as for \code{osm_input} but the input should provide symplastic relative water content as well and as a result the output includes the slope and intercept for both the bulk and symplastic variables.
}
}
