% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sumParams.R
\name{sumParams}
\alias{sumParams}
\title{Summarize estimated pressure volume curve parameters by group}
\usage{
sumParams(
  df,
  group_name,
  remove.cols = F,
  cols.to.remove = c("leaf", "unique_id")
)
}
\arguments{
\item{df}{A data frame. Data frame containing predicted parameters.}

\item{group_name}{A string. Name of variable to group summary by.}

\item{remove.cols}{T/F. Should columns be removed in the output}

\item{cols.to.remove}{A list. List of column names to be removed from using select. Default is leaf and unique_id}
}
\value{
Returns a summarized data frame with the mean parameter values.
}
\description{
Summarize estimated pressure volume curve parameters by group
}
