% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osmotic_estimates.R
\name{sym_rwc}
\alias{sym_rwc}
\title{Estimate symplastic water content}
\usage{
sym_rwc(rwd, sma_mod)
}
\arguments{
\item{sma_mod}{An object of class "sma_model" containing the slope and intercept of the linear relationship between relative water deficit and negative inverse water potential.}

\item{rwc}{Vector of relative water content values.}
}
\value{
A vector of symplastic relative water content values.
}
\description{
Estimate symplastic water content
}
