% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tlp_estimates.R
\name{tlpinput}
\alias{tlpinput}
\title{Constructor for the tlp_input class}
\usage{
tlpinput(
  psi_above,
  psip,
  rwd,
  symrwd,
  rwc_above,
  symrwc_above,
  psi_below,
  rwc_below,
  symrwc_below
)
}
\description{
Constructor for the tlp_input class
}
